Day06 Class

+++ src/day06_ArithmeticOperators/KgToPounds.java
@@ -4,7 +4,7 @@ public class KgToPounds {
 
     public static void main(String[] args) {
 
-        int kg = 100;
+        double kg = 100; // 100.0
         double pound= kg * 2.20462;
 
         System.out.println(kg+" kg equal to "+ pound+" pounds");
